Big features update
I've added several features, including a nav bar on the month page, to navigate to other months. I've also added a nav bar at the top for getting 'home' and for an 'about' page. I've also added a link to a search page that I could add in later.

I've added a banner and month selector on the home page too.